pyproject.toml
[project]
name = "webvtt-loc"
version = "0.2.0"
requires-python = ">=3.11"
dependencies = [
    "argparse>=1.4.0",
    "webvtt-py>=0.5.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
